made one lev_controllers file with one trait for each controller
