+++ lib/control/src/lib.rs
@@ -1,6 +1,6 @@
 #![cfg_attr(not(test), no_std)]
 
 pub mod hp_relay;
-pub mod pi_controller;
+pub mod lev_controllers;
 pub mod pid_controller;
 pub mod pneumatics;
